requirements.txt
pyspark==4.1.1
apache-airflow==3.1.7
prometheus_client==0.24.1
pyyaml
pandas
numpy